Life table

Life table

Usage

LT(age, sex = "m", mx, ax = NULL, w = NULL, l0 = 1)

Arguments

age

Numeric array of age intervals; for full life table = 0:100; for concise life table = c(0:1, seq(5,85,5))

sex

Character. Sex. "m" for males or "f" for females. By default = "m".

mx

Numeric array with age specific mortality rates.

ax

Optional. Numeric array with ax. By default, it is the middle of the interval, while ax for age 0 is modeled as in Andreev & Kingkade (2015).

w

Optional. Numeric array with weights for each age interval for calculating weighted life expectancy (wex).

l0

Numeric. Life table radix. By default, = 1 but it can be any positive real number. In "human" demography tradition it is 100'000, in "ecological" and "evolutionary" demography tradition it is 1.

Value

Matrix of (age x 9). Columns are: age, mx, ax, qx, lx, dx, Lx, Tx, ex

Details

By default, \(a_x\) for age 0 (first entity in ax) is modeled as in Andreev & Kingkade (2015, p. 390, see table 3-2).

The weighted life expectancy is calculated as follows: $$e_x^w = \frac{\sum_{i = x}^{\omega}L_x w_x}{l_x}$$ where \(\omega\) is the last age, \(w\) is weight s.t. \(w \in [0,1]\), and other variables are life table functions.

References

Andreev, E. M., & Kingkade, W. W. (2015). Average age at death in infancy and infant mortality level: Reconsidering the Coale-Demeny formulas at current levels of low mortality. Demographic Research, 33, 363-390.

See also