Cohort Component Model for Projecting Population

Cohort Component Model for Projecting Population

Usage

ccm(Mx, Fx, Ix = NULL, age.mx, age.fx, N0, ...)

Arguments

Mx

Matrix or dataframe with mx, where row represents age, column represents a forecasted period. Thus, first column should be the first projected period, while each row represents age-specific mortality rate.

Fx

Matrix or dataframe with fx, where row represents age, column represents a forecasted period. Thus, first column should be the first projected period, while each row represents age-specific fertility rate.

Ix

Optional. Matrix or dataframe with net number of migrants, where row represents age, column represents a forecasted period. Thus, first column should be the first projected period, while each row represents net number of migrants.

age.mx

Numeric vector. Age groups for Mx rows.

age.fx

Numeric vector. Age groups for Fx rows.

N0

Numeric vector. Population in the initial period by age groups.

...

Optional. Additional arguments for leslie function.

Value

Matrix with projected population. Columns represent periods, rows represent age groups. The first column is N0 (initial period).